fixed from code review
Dan H’s code review:

https://github.com/NextITCorp/pyte/blob/master/pyte/DateExtensions.py#L1
6

This assumes that the caller is also using pytz for their zone-aware
dates, but pytz isn't standard and code involving other libraries may
use a different UTC object (which sadly won't compare equal to
pytz.utc). This should work with any reasonable implementation of a UTC
timezone:

    return theDateTime.tzname() == 'UTC'

Using a string compare for this may seem like a hack, but I think it's
a lesser evil. Alternatively you could test `theDateTime.utcoffset() ==
timedelta(0)`, but that probably runs afoul of some edge case involving
daylight savings time.

https://github.com/NextITCorp/pyte/blob/master/pyte/DateExtensions.py#L2
6

Probably wanted `or` here?

https://github.com/NextITCorp/pyte/blob/master/pyte/DateExtensions.py#L2
9

Not sure which Python versions you're targeting, but this won't work in
2.x - timedelta/timedelta division was added in 3.2. You can use the
`days` property of the timedelta object instead, which also rules out
any floating-point shenanigans:

    for dist in range((endDate - startDate).days):

https://github.com/NextITCorp/pyte/blob/master/pyte/DateExtensions.py#L4
0

Style nit: The list constructor is a more direct way to collect the
elements of an iterable, no dummy binding required:

    return list(dateRangeGenerator(startDate, endDate))

https://github.com/NextITCorp/pyte/blob/master/pyte/TemporalExpressions.
py#L17

Obligatory "tsk, don't use import * in non-throwaway code" remark.

https://github.com/NextITCorp/pyte/blob/master/pyte/TemporalExpressions.
py#L62

Duplicates the body of `dayMatches`.

https://github.com/NextITCorp/pyte/blob/master/pyte/TemporalExpressions.
py#L82-L91

Style nits:
- Instead of `x > lo and x < hi`, you can write `lo < x < hi`. Just
like in math!
- The mixed ands and ors at the same indentation level makes that
conditional really hard to parse visually. Since it's really a
top-level `or`, I'd line those up with the edge of the `return`, and if
you want more line-splitting for the `and`s then indent them so the
overall structure is clearer. Right-aligning boolean operators beneath
the return keyword is just my preferred idiom, it's not officially
blessed PEP-8 stuff or anything. But it sure is easy on the eyes:

    if startMonth == endMonth:
        return candidate.month == startMonth \
           and startDay <= candidate.day <= endDay

    return startMonth < candidate.month < endMonth \
        or candidate.month == startMonth and candidate.day >= startDay \
        or candidate.month == endMonth and candidate.day <= endDay

And a semantic nit: rangeInYear works with an inclusive range, instead
of the typical half-open ranges Python uses (your dateRange helper is
also half-open). Trying to remember which functions use which range
convention is a pain, so I think half-open everything is the way to go.
But if you've got a compelling reason to make this one inclusive, I
suggest consistently using inclusive date ranges everywhere in the
library.

https://github.com/NextITCorp/pyte/blob/master/pyte/TemporalExpressions.
py#L105-L106

When you're building up a sequence just to immediately iterate it, it's
a good habit to use a generator expression instead of a list
comprehension (switch the square brackets to parens), so you don't
realize the entire list in memory at once. But for the particular case
of boolean reductions there are handy `any` and `all` built-ins:

    return any(tefunc(candidate) for tefunc in teList)

Note that this version short-circuits on the first True result, so if
for some reason you're evaluating those predicates for side-effects
you'd need to stick with reduce, or just use `any` with a list
comprehension.

It's also worth mentioning that `operator.or_` and `operator.and_`
aren't the droids you're looking for:

    >>> operator.or_(1, 2)
    3

Surprise! It's actually bitwise or.

https://github.com/NextITCorp/pyte/blob/master/pyte/TemporalExpressions.
py#L120-L121

Ditto, except s/any/all/.

https://github.com/NextITCorp/pyte/blob/master/pyte/TemporalExpressions.
py#L161-L164

Python belongs to the class of sane languages that have a separate
operator for integer division, so `int(x/y)` can usually be written
more clearly as `x//y`. ("Usually" because if your inputs are floats,
// will return a floored float. But you're dealing with ints here.)
Confidentiality Notice: This e-mail and any attachments may contain
confidential material for the sole use of the intended recipient(s).
Any review or distribution by others is strictly prohibited. If you are
not the intended recipient, please contact the sender and delete all
copies.